requirements.txt
matplotlib~=3.1.2
numpy~=1.17.4
pandas~=1.3.4
tensorflow~=2.7.0
tqdm~=4.62.3
scikit-learn~=1.0.1
scipy~=1.7.3
keras~=2.7.0
mayavi~=4.7.4
requests~=2.22.0
gstlearn~=0.1.27
seaborn~=0.11.2
deel.lip